Move network wrapper to own file
